Enable qodana annotations for PRs from forks.
This PR moves the annotation part from the `tests.yml` workflow
to a new workflow. This is done for security reasons, the `tests.yml`
workflow works on untrusted user-input. That's why it only produces
an artifact that is stored. The new workflow then ONLY has permissions
to use the checks API (used for annotations). It downloads the artifact
from the `tests.yml` workflow  and is limited to performing the
annotation based on this untrusted user-input. This is the only usage
of user-input in `annotate-qodana-results.yml` besides getting the
PR number, for which we check that it is indeed a number.